Attempt to fix A* path search by recalculating heuristic values of nodes properly.

+++ src/viewport_func.h
@@ -18,7 +18,7 @@
 
 #include <unordered_set>
 
-typedef uint16 PathCost;
+typedef float PathCost;
 
 extern TileIndex ship_planner_start_tile;
 extern TileIndex ship_planner_end_tile;
